feat(product-service): split handlers
